[Agent] extract TestData into separate module
