+++ tests/common/entities/index.js
@@ -1,6 +1,7 @@
 export { default as SimpleEntityManager } from './simpleEntityManager.js';
 export { default as TestBed } from './testBed.js';
 export * from './testBed.js';
+export { TestData } from './testData.js';
 export * from './serializationUtils.js';
 export * from './execContext.js';
 export * from './invalidInputHelpers.js';
